requirements.txt
streamlit==1.28.1
langchain==0.0.350
langchain-community==0.0.10
pypdf2==3.0.1
python-docx==1.1.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
transformers==4.40.0
torch==2.1.1
huggingface_hub==0.23.0